Major refactor of fetchContent
