+++ src/UrlProcessor.ts
@@ -1,7 +1,11 @@
+import type {MetaData} from './types';
+
+
 import {fixDoubleSlashes, getProjectLocaleConfigById, RENDER_MODE} from './utils';
 import {commonChars} from './utils/commonChars';
-import {ImageData, LinkData, MetaData} from './guillotine/getMetaData';
+import {ImageData, LinkData} from './guillotine/getMetaData';
 import {addBasePath} from 'next/dist/client/add-base-path';
+
 
 export class UrlProcessor {
