+++ src/guillotine/getCleanContentPathArrayOrThrow400.ts
@@ -1,0 +1,22 @@
+/** Checks a site-relative contentPath as a slash-delimited string or a string array, and returns a pure site-relative path string (no double slashes, starts with a slash, does not end with one). */
+export const getCleanContentPathArrayOrThrow400 = (contentPath: string | string[] | undefined): string => {
+    if (contentPath === undefined) {
+        return '';
+    }
+    const isArray = Array.isArray(contentPath);
+
+    if (!isArray) {
+        if (typeof contentPath !== 'string') {
+            throw Error(JSON.stringify({
+                code: 400,
+                message: `Unexpected target content _path: contentPath must be a string or pure string array (contentPath=${JSON.stringify(
+                    contentPath)})`,
+            }));
+        }
+
+        return contentPath;
+
+    } else {
+        return (contentPath).join('/');
+    }
+};
