+++ src/index.ts
@@ -1,4 +1,15 @@
-export type {Context} from './types'; // tsc doesn't copy d.ts files to dist
+export type {
+    ComponentDefinition,
+    ContentFetcher,
+    Context,
+    FetchContentResult,
+    MetaData,
+    PageComponent,
+    PageData,
+    PageRegion,
+    RegionTree,
+    SelectedQueryMaybeVariablesFunc
+} from './types'; // tsc doesn't copy d.ts files to dist
 
 export {XP_BASE_URL_HEADER} from './constants';
 export * from './utils';
