+++ test/guillotine/fetchContent.test.ts
@@ -1,0 +1,342 @@
+import type {
+    ContentApiBaseBody,
+    Context,
+    ProjectLocaleConfig
+} from '../../src/types';
+
+import {
+    // afterAll,
+    // beforeAll,
+    beforeEach,
+    describe,
+    expect,
+    jest,
+    test as it
+} from '@jest/globals';
+import {XP_BASE_URL_HEADER} from '../../src/constants';
+import { afterEach } from 'node:test';
+// import {ENONIC_APP_NAME} from '../constants';
+
+
+global.console = {
+    // error: console.error,
+    error: jest.fn(),
+    warn: jest.fn(),
+    log: jest.fn(),
+    info: jest.fn(),
+    debug: jest.fn(),
+} as unknown as Console;
+
+
+const ENONIC_APP_NAME = 'com.enonic.web.enonic.com';
+
+const FETCH_FROM_API_PARAMS_VALID: [
+    string,
+    ContentApiBaseBody,
+    ProjectLocaleConfig,
+    {}
+] = [
+    'http://localhost:8080/site/enonic-homepage/master',
+    {
+        query: `{
+    guillotine {
+        query(first: 1) {
+            valid
+        }
+    }
+}`,
+        variables: {
+            // key: 'value'
+        }
+    },
+    {
+        default: true,
+        project: 'enonic-homepage',
+        site: 'enonic-homepage',
+        locale: 'en'
+    },
+    {
+        // headerKey: 'headerValue'
+    }
+];
+
+const OLD_ENV = process.env;
+
+
+describe('guillotine', () => {
+    // beforeAll(() => {});
+
+    beforeEach(() => {
+        jest.replaceProperty(process, 'env', {
+            ...OLD_ENV,
+            // ENONIC_API_TOKEN: '1234567890',
+            ENONIC_APP_NAME,
+            ENONIC_PROJECTS: 'en:enonic-homepage/enonic-homepage'
+        });
+        jest.spyOn(global, 'fetch').mockImplementation((input, init = {}) => {
+            // console.debug('fetch', input, init);
+            const guillotine = {
+                query: [{
+                    valid: true
+                }]
+            };
+            return Promise.resolve({
+                json: () => Promise.resolve({
+                    data: {
+                        guillotine
+                    }
+                }),
+                text: () => Promise.resolve(JSON.stringify({
+                    guillotine
+                })),
+                ok: true,
+                status: 200
+            } as Response);
+        });
+    }); // beforeEach
+
+    afterEach(() => {
+        jest.resetModules();
+        jest.restoreAllMocks(); // Restores all mocks and replaced properties back to their original value.
+    });
+
+    // afterAll(() => {});
+
+    // describe('fetchContent', () => {
+    //     const RENDER_MODES = ['edit','inline','preview','live','admin','next'];
+    //     RENDER_MODES.forEach((mode) => {
+    //         it(`${mode}`, () => {
+    //             import('../../src/guillotine/fetchContent').then((moduleName) => {
+    //                 const context: Context = {
+    //                     headers: {
+    //                         get(name: string) {
+    //                             if (name === 'content-studio-mode') {
+    //                                 return mode;
+    //                             }
+    //                             if (name === 'content-studio-project') {
+    //                                 return 'prosjekt';
+    //                             }
+    //                             if (name === 'jsessionid') {
+    //                                 return '1234';
+    //                             }
+    //                             if (name === XP_BASE_URL_HEADER) {
+    //                                 return '/whatnot/edit/1234';
+    //                             }
+    //                             console.error('headers get name', name);
+    //                         }
+    //                     },
+    //                 } as Context;
+    //                 expect(moduleName.fetchContent(context)).toStrictEqual({});
+    //             });
+    //         });
+    //     });
+    // });
+
+    describe('fetchFromApi', () => {
+        it('returns query results', () => {
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                moduleName.fetchFromApi(...FETCH_FROM_API_PARAMS_VALID).then((result) => {
+                    console.debug(result);
+                    expect(result).toEqual({
+                        data: {
+                            guillotine: {
+                                query: [{
+                                    valid: true
+                                }]
+                            }
+                        }
+                    });
+                });
+            });
+        });
+
+        it('throws a nice error when fetch throws', () => {
+            jest.spyOn(global, 'fetch').mockImplementation(() => {
+                throw new Error('fetch error');
+            });
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                expect(() => moduleName.fetchFromApi(...FETCH_FROM_API_PARAMS_VALID)).rejects.toThrow(Error(JSON.stringify({
+                    code: 'API',
+                    message: 'fetch error'
+                })));
+            });
+        });
+    }); // fetchFromApi
+
+    describe('fetchGuillotine', () => {
+        it('fetches a response from guillotine', () => {
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                moduleName.fetchGuillotine(...FETCH_FROM_API_PARAMS_VALID).then((result) => {
+                    expect(result).toEqual({
+                        guillotine: {
+                            query: [{
+                                valid: true
+                            }]
+                        }
+                    });
+                });
+            });
+        });
+
+        it('returns an error when query is not a string', () => {
+            const [contentApiUrl, body, projectConfig, headers] = [
+                FETCH_FROM_API_PARAMS_VALID[0],
+                {
+                    query: true as unknown as string,
+                    variables: {}
+                },
+                FETCH_FROM_API_PARAMS_VALID[2],
+                FETCH_FROM_API_PARAMS_VALID[3]
+            ]
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                moduleName.fetchGuillotine(contentApiUrl, body, projectConfig, headers).then((result) => {
+                    expect(result).toEqual({
+                        error: {
+                            code: '400',
+                            message: 'Invalid or missing query. JSON.stringify(query) = true'
+                        }
+                    });
+                });
+            });
+        });
+
+        it('returns an error when query is an empty string', () => {
+            const [contentApiUrl, body, projectConfig, headers] = [
+                FETCH_FROM_API_PARAMS_VALID[0],
+                {
+                    query: '',
+                    variables: {}
+                },
+                FETCH_FROM_API_PARAMS_VALID[2],
+                FETCH_FROM_API_PARAMS_VALID[3],
+            ]
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                moduleName.fetchGuillotine(contentApiUrl, body, projectConfig, headers).then((result) => {
+                    expect(result).toEqual({
+                        error: {
+                            code: '400',
+                            message: 'Invalid or missing query. JSON.stringify(query) = ""'
+                        }
+                    });
+                });
+            });
+        });
+
+        it('ensures json.errors is always an array', () => {
+            jest.spyOn(global, 'fetch').mockImplementation((input, init = {}) => {
+                const json = {
+                    errors: 'Single string error'
+                };
+                return Promise.resolve({
+                    json: () => Promise.resolve(json),
+                    text: () => Promise.resolve(JSON.stringify(json)),
+                    ok: true,
+                    status: 200
+                } as Response);
+            });
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                moduleName.fetchGuillotine(...FETCH_FROM_API_PARAMS_VALID).then((result) => {
+                    expect(result).toEqual({
+                        error: {
+                            code: '500',
+                            message: 'Server responded with 1 error(s), probably from guillotine - see log.'
+                        }
+                    });
+                });
+            });
+        });
+
+        it('returns a nice error when fetch throws', () => {
+            jest.spyOn(global, 'fetch').mockImplementation(() => {
+                throw new Error('fetch error');
+            });
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                expect(moduleName.fetchGuillotine(...FETCH_FROM_API_PARAMS_VALID)).resolves.toEqual({
+                    error: {
+                        code: 'API',
+                        message: 'fetch error'
+                    }
+                });
+            });
+        });
+
+        it('returns a nice error when fetch response not ok', () => {
+            jest.spyOn(global, 'fetch').mockImplementation(() => {
+                return Promise.resolve({
+                    json: () => Promise.resolve({}),
+                    text: () => Promise.resolve('Internal server error'),
+                    ok: false,
+                    status: 500
+                } as Response);
+            });
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                expect(moduleName.fetchGuillotine(...FETCH_FROM_API_PARAMS_VALID)).resolves.toEqual({
+                    error: {
+                        code: 500,
+                        message: "Data fetching failed (message: 'Internal server error')"
+                    }
+                });
+            });
+        });
+
+        it('returns a nice error when fetch response is not json', () => {
+            jest.spyOn(global, 'fetch').mockImplementation(() => {
+                return Promise.resolve({
+                    text: () => Promise.resolve('Could this ever happen? I guess if endpoint url is wrong?'),
+                    ok: true,
+                    status: 200
+                } as Response);
+            });
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                expect(moduleName.fetchGuillotine(...FETCH_FROM_API_PARAMS_VALID)).resolves.toEqual({
+                    error: {
+                        code: 500,
+                        message: 'API call completed but with non-JSON data: "Could this ever happen? I guess if endpoint url is wrong?"'
+                    }
+                });
+            });
+        });
+
+        it('returns a nice error when fetch response is empty', () => {
+            jest.spyOn(global, 'fetch').mockImplementation(() => {
+                return Promise.resolve({
+                    json: () => Promise.resolve(),
+                    text: () => Promise.resolve('Could this ever happen?'),
+                    ok: true,
+                    status: 200
+                } as Response);
+            });
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                expect(moduleName.fetchGuillotine(...FETCH_FROM_API_PARAMS_VALID)).resolves.toEqual({
+                    error: {
+                        code: 500,
+                        message: 'API call completed but with unexpectedly empty data: "Could this ever happen?"'
+                    }
+                });
+            });
+        });
+
+        it("returns a nice error when error.message can't be JSON parsed", () => {
+            import('../../src/guillotine/fetchContent').then((moduleName) => {
+                const spyOnFetchFromApi = jest.spyOn(moduleName, 'fetchFromApi').mockImplementation(async () => {
+                    throw new Error('not json parseable');
+                });
+                expect(moduleName.fetchGuillotine(...FETCH_FROM_API_PARAMS_VALID)).resolves.toEqual({
+                    error: {
+                        code: 'Client-side error',
+                        message: 'not json parseable'
+                    }
+                });
+                expect(spyOnFetchFromApi).toHaveBeenCalledTimes(1);
+            });
+        });
+    }); // fetchGuillotine
+
+    // describe('fetchMetaData', () => {
+    //     import('../../src/guillotine/fetchContent').then((moduleName) => {
+    //         it('', () => {
+    //             expect(moduleName.fetchMetaData()).toStrictEqual({});
+    //         });
+    //     });
+    // });
+});
